fix: revert emailjs patch & avoid its types
Patches cannot be used inside of libraries.

We don't actually need the fs import from emailjs that requires es module interop,
since we are suggesting that attachments provide the `data` directly.
I'm just avoiding importing it from TS's perspective to avoid that error.
I've got a PR open with the library to hopefully resolve it.